move handleTerminalAttach to generic build
